chore(py3): move the ball forward on py3

+++ gdc_client/query/index.py
@@ -1,6 +1,6 @@
 import logging
 from json import dumps
-from urlparse import urljoin
+from six.moves.urllib.parse import urljoin
 
 import requests
 
@@ -34,12 +34,12 @@ class GDCIndexClient(object):
             return self.metadata[uuid]['md5sum']
 
     def get_filesize(self, uuid):
-        # type: (str) -> long
+        # type: (str) -> int
         if uuid in self.metadata.keys():
-            return long(self.metadata[uuid]['file_size'])
+            return int(self.metadata[uuid]['file_size'])
 
     def get_access(self, uuid):
-        # type: (str) -> long
+        import pdb; pdb.set_trace()
         if uuid in self.metadata.keys():
             return self.metadata[uuid]['access']
